requirements.txt
flask==1.1.2
pyyaml==5.4.1